telegram_bot/requirements.txt
aiogram>=3.13.0
apscheduler>=3.10.4
aiohttp>=3.10.11
asyncpg>=0.30.0
SQLAlchemy>=2.0.36
openai>=1.0.0
bitrix24-rest>=2.0.0
python-dotenv>=1.0.1
alembic>=1.13.1
# Для семантического поиска будем использовать легковесные библиотеки
numpy>=1.26.4 
scikit-learn>=1.4.2
openai>=1.0.0 # Уже есть, но оставляем
tiktoken>=0.6.0
pydantic>=2.7.1
loguru>=0.7.2
gspread>=6.1.2
google-auth>=2.34.0
